+++ sas-20181203/core/Models/DescribeOnceTaskRequest.cs
@@ -52,6 +52,10 @@ namespace AlibabaCloud.SDK.Sas20181203.Models
         [Validation(Required=false)]
         public string RootTaskId { get; set; }
 
+        [NameInMap("Source")]
+        [Validation(Required=false)]
+        public string Source { get; set; }
+
         /// <summary>
         /// <para>The timestamp when the root task starts. Unit: milliseconds.</para>
         ///